+++ core/utils.py
@@ -1,0 +1,12 @@
+def format_currency(value):
+    """통화 값을 읽기 쉬운 형식으로 변환합니다."""
+    if value > 1_000_000_000_000:
+        value = f"{value/1_000_000_000_000:,.2f}T"
+    elif value > 1_000_000_000:
+        value = f"{value/1_000_000_000:,.2f}B"
+    elif value > 1_000_000:
+        value = f"{value/1_000_000:,.2f}M"
+    elif value > 1_000:
+        value = f"{value/1_000:,.2f}K"
+
+    return value
